<refactor>[front] refactor LoadMorePost and its children
- Clearly divide the roles of each component.
    - <PostListItem/> only show post card
    - <loadMorePosts/> load posts and configure animation of loading
 - refactor PostMenu
