+++ front/src/components/posts/view/postMenu.tsx
@@ -9,9 +9,13 @@ type PostMenuProps = {
     post : PostType
 }
 
+/*TODO
+- Post Owner만 삭제하도록 로직 수정 
+*/
 
 const PostMenu: React.FC<PostMenuProps> = ({post}) => {
-    const bg = customColors.black[300];
+    const normalColor = customColors.black[100];
+    const focusColor = customColors.purple[100];
     const isOwner = false;
 
     return (
@@ -19,14 +23,15 @@ const PostMenu: React.FC<PostMenuProps> = ({post}) => {
             <MenuButton
                 as={IconButton}
                 aria-label="Options"
-                icon={<HamburgerIcon/>}
-                bg={bg}
+                icon={<HamburgerIcon w="24px" h="24px"/>}
                 border={"none"}
+                variant="outline"
+                color = {normalColor}
+
             />
             <MenuList>
                 <MenuItem>Report</MenuItem>
                 <MenuItem color={customColors.error[100]}>Delete</MenuItem>
-                
             </MenuList>
         </Menu>
     )
